made client able to run forever unattended
client now has -f or --forever flag for making it reconnected
endlessly after each termination, or upon error. so now many
clients can be started and then the server iterated upon without
stopping to turn the clients on over and over!

the client also now waits quietly for connect to work, ignoring
errors. not sure if that's the best way to handle that, but...